refactor: split rest bootstrap anchors
