+++ .sampo/changesets/1012-split-rest-bootstrap-anchors.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split REST workspace bootstrap anchor patching from sync-rest script patching while preserving compatibility exports.
