+++ packages/wp-typia-project-tools/src/runtime/cli-add-workspace-rest-generated.ts
@@ -15,6 +15,8 @@ import {
 import {
 	ensureRestResourceBootstrapAnchors,
 	ensureRestSchemaHelperBootstrapAnchors,
+} from "./cli-add-workspace-rest-bootstrap-anchors.js";
+import {
 	ensureRestResourceSyncScriptAnchors,
 } from "./cli-add-workspace-rest-anchors.js";
 import {
